+++ mastermlx/robotics/__init__.py
@@ -1,0 +1,52 @@
+"""Robotics primitives built on top of NumPy."""
+
+from .transforms import (
+    compose_transform,
+    euler_to_matrix,
+    homogeneous_transform,
+    invert_transform,
+    matrix_to_euler,
+    matrix_to_quaternion,
+    quaternion_to_matrix,
+    rpy_to_matrix,
+    rot_x,
+    rot_y,
+    rot_z,
+    skew,
+    transform_points,
+    unskew,
+)
+from .kinematics import DHLink, chain_positions, dh_transform, forward_kinematics, inverse_kinematics
+from .jacobian import finite_difference_jacobian, geometric_jacobian, planar_2r_jacobian
+from .trajectory import cubic_time_scaling, joint_trajectory, quintic_time_scaling, sample_joint_trajectory
+from .visualizer import plot_chain
+
+__all__ = [
+    "DHLink",
+    "chain_positions",
+    "compose_transform",
+    "cubic_time_scaling",
+    "dh_transform",
+    "euler_to_matrix",
+    "finite_difference_jacobian",
+    "forward_kinematics",
+    "geometric_jacobian",
+    "homogeneous_transform",
+    "invert_transform",
+    "inverse_kinematics",
+    "joint_trajectory",
+    "matrix_to_euler",
+    "matrix_to_quaternion",
+    "planar_2r_jacobian",
+    "plot_chain",
+    "quaternion_to_matrix",
+    "quintic_time_scaling",
+    "rpy_to_matrix",
+    "rot_x",
+    "rot_y",
+    "rot_z",
+    "sample_joint_trajectory",
+    "skew",
+    "transform_points",
+    "unskew",
+]